bot/plugins/nonebot-plugin-kuraku-help/pyproject.toml
[project]
name = "nonebot-plugin-kuraku-help"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "ssttkkl", email = "huang.wen.long@hotmail.com"},
]
dependencies = [
    "nonebot2<3.0.0,>=2.3.0",
    "nonebot-plugin-send-anything-anywhere>=0.3.0",
    "nonebot-plugin-access-control-api",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
